requirements.txt
langchain
openai
azure-ai-inference
python-dotenv
pytest
pytest-mock
requests
PyGithub